requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.7.0
boto3==1.29.0
botocore==1.32.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
decorator==5.1.1
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
executing==2.0.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-DebugToolbar @ git+https://github.com/pallets-eco/flask-debugtoolbar@ce02d2da3c37b3212eebafa9392c429b91d13923
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
forex-python==1.8
greenlet==3.0.1
idna==3.4
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments @ file:///private/tmp/pygments-20230806-6216-4zajry/Pygments-2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
simplejson==3.19.1
six==1.16.0
SQLAlchemy==2.0.23
stack-data==0.6.3
traitlets==5.11.2
typing_extensions==4.8.0
urllib3==2.0.5
wcwidth==0.2.8
Werkzeug==3.0.1
WTForms==3.1.1
